Remove test for local version matching
There is only one known use case that you would care that the local
version is different than the pkg_resource distribution version. This
would only occur if you were in the pylmod directory and using python
from there. In this case you would get the globally installed version
instead of the checked out version after this commit.

On the other hand, readthedocs.org does exactly this (and gets the
wrong version), and it is much more likely that we would forget to
update the version in multiple places (i.e. docs/conf.py and setup.py)
so we have chosen to accept the narrower error case.