Add get_staff() and get_options() to GradeBook.
get_staff() returns staff membership for a gradebook. The data
returned is de-normalized from what is returned by the LMod Web
service. Due to the comprehensive amount of data returned, a 'simple'
parameter is available to return less data.

get_options() returns meta-data about the gradebook. It is part
of this PR to help map courses to membership groups. One such
group is the staff list.

Partially addresses #23
Closes Issue #41